move code from abstract classes to normal
