+++ src/userModule/interface/auth.js
@@ -4,13 +4,14 @@ const Joi = require('joi');
 const express = require('express');
 const router = express.Router();
 const moongose = require('mongoose');
-const _=require('lodash');
+const _= require ('lodash');
 //const User = require ('...UserRepository')
 
 router.post('/', async (req, res) =>
 {
     const { error } = validate(req.body);
     if (error) return res.status(400).send(error.details[0].message);
+
 
     let user = await User.findOne({ email: req.body.email });
     if (!user) return res.status(400).send('Invalid email or pasword.');
