+++ src/userModule/interface/hash.js
